Change "Under Construction" watermark (#40)
Change "UNDER CONSTRUCTION - DO NOT USE" to "UNDER CONSTRUCTION"

In my opinion "do not use" is to restrictive and lets (new) user ignoe
the site. So they do not even report errors / missing info or even
contribute content. I think that a warning that the info is incomplete
is OK and sufficient.

Please discuss in docu group and accept or cancel PR.